Cleaned up
